more class shifting to their own specific files. Rename funcheadertype to a more appropriate className: 'FunctionHeader'
