+++ src/include/type.hpp
@@ -136,22 +136,5 @@ struct exprtype
 
 void report(njnr::logType t, std::string s);
 
-/* function prototype data structure */
-class funcheadertype
-{
-   public:
-   funcheadertype() : name{""},
-                      paramlist{nullptr},
-                      returntype{njnr::type::VOID},
-                      ttype{njnr::type::VOID}
-                      {}
-
-   virtual ~funcheadertype();
-   const std::string toString() const;
-   std::string   name;        // fn name
-   std::shared_ptr<njnr::List>   paramlist;   // list of parameters the fn accept as input
-   njnr::type    returntype;  // return type(if any) of fn
-   njnr::type    ttype;       // type of fn?(TBD)
-};
 }  // namespace njnr
 #endif  // SRC_INCLUDE_TYPE_HPP_
